style: mirror lib hierarchy
