+++ types/passport-saml-metadata/src/metadata.d.ts
@@ -1,0 +1,3 @@
+import { SamlConfig } from "passport-saml";
+
+export function metadata(config: SamlConfig): (() => void);
